Refactoring: Moving BiometricEncryption to HeimdallEncryptionManager
